feat: intial code to download node based on operating system  and architecture

+++ src-build/downloadNode.js
@@ -1,5 +1,16 @@
 import * as https from "https";
 import * as fs from "fs";
+import tar from "tar";
+import * as path from "path";
+import AdmZip from 'adm-zip';
+import {fileURLToPath} from 'url';
+import {dirname} from 'path';
+import * as fsExtra from "fs-extra";
+
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
 const LTS_URL_PREFIX = 'https://nodejs.org/dist/latest-v18.x/';
 
 /**
@@ -38,24 +49,29 @@ async function fetchLatestNodeVersion() {
  */
 
 async function downloadNodeBinary(version, platform, arch) {
-    const MAX_RETRIES = 3;
     const extension = (platform === "win") ? "zip" : "tar.gz";
     const fileName = `node-v${version}-${platform}-${arch}.${extension}`;
 
+    const fullPath = `${__dirname}/${fileName}`
+    // Check if the file already exists
+    if (fs.existsSync(fullPath)) {
+        console.log(`File ${fileName} already exists. No need to download.`);
+        return fileName;
+    }
+    const MAX_RETRIES = 3
     for (let attempt = 0; attempt < MAX_RETRIES; attempt++) {
         try {
-            const file = fs.createWriteStream(fileName);
+            const file = fs.createWriteStream(fullPath);
             await new Promise((resolve, reject) => {
-                https.get(`${LTS_URL_PREFIX}node-v${version}-${platform}-${arch}.${extension}`,
-                    (res) => {
-                        res.pipe(file);
-                        res.on('end', () => resolve(fileName));
-                        res.on('error', (err) => {
-                            fs.unlink(fileName, () => {
-                            }); // Remove the file on error
-                            reject(err);
-                        });
+                https.get(`${LTS_URL_PREFIX}node-v${version}-${platform}-${arch}.${extension}`, (res) => {
+                    res.pipe(file);
+                    res.on('end', () => resolve(fileName));
+                    res.on('error', (err) => {
+                        fs.unlink(fileName, () => {
+                        }); // Remove the file on error
+                        reject(err);
                     });
+                });
             });
             return fileName; // If the download was successful, return the file name
         } catch (err) {
@@ -69,10 +85,131 @@ async function downloadNodeBinary(version, platform, arch) {
 }
 
 
-try {
-    const version = await fetchLatestNodeVersion();
-    const fileName = await downloadNodeBinary(version, 'linux', 'x64');
-    console.log('Downloaded file:', fileName);
-} catch (err) {
-    console.error(err);
+/**
+ * Extracts a tar archive file to the specified output directory.
+ * @param {string} inputFile - The path of the tar archive file to be extracted.
+ * @param {string} outputDir - The path of the directory where the files will be extracted.
+ * @returns {Promise<void>} - A Promise that resolves when the extraction is complete.
+ */
+async function untarFile(inputFile, outputDir) {
+    // Ensure that inputFile and outputDir are absolute paths
+    const file = path.resolve(inputFile);
+    const outdir = path.resolve(outputDir);
+
+    try {
+        await tar.x({
+            file: file,
+            cwd: outdir,
+        });
+        console.log('Extraction complete');
+    } catch (err) {
+        console.error(err);
+    }
 }
+
+/**
+ Unzips a file at the specified path to the specified extraction path.
+ @param {string} zipFilePath - The path to the ZIP file to be extracted.
+ @param {string} extractPath - The path where the contents of the ZIP file should be extracted.
+ @returns {void}
+ */
+function unzipFile(zipFilePath, extractPath) {
+    try {
+        let zip = new AdmZip(zipFilePath);
+        zip.extractAllTo(/*target path*/extractPath, /*overwrite*/true);
+        console.log(`File has been unzipped to ${extractPath}`);
+    } catch (err) {
+        console.error(err);
+    }
+}
+/**
+ * Copies the latest version of Node.js binary for a specific platform and architecture.
+ * @param {string} platform - The platform for which to download the Node.js binary. (e.g., "win", "linux", "mac")
+ * @param {string} arch - The architecture for which to download the Node.js binary. (e.g., "x86", "x64")
+ * @returns {Promise<void>} - A Promise that resolves when the Node.js binary is copied successfully.
+ */
+async function copyLatestNodeForBuild(platform, arch) {
+    const version = await fetchLatestNodeVersion();
+    const fileName = await downloadNodeBinary(version, platform, arch);
+    const fullPath = `${__dirname}/${fileName}`
+    let nodeFolder = "";
+
+    if (platform === "win") {
+        unzipFile(fullPath, ".");
+        nodeFolder = fileName.slice(0, -4);
+    } else {
+        await untarFile(fullPath, ".");
+        nodeFolder = fileName.slice(0, -7);
+    }
+    console.log(nodeFolder);
+    const fullPathUnzipFolder = `${__dirname}/${nodeFolder}`;
+    const fullPathOfNode = `${__dirname}/node`;
+    await removeDir(fullPathOfNode);
+
+    try {
+        fs.renameSync(fullPathUnzipFolder, fullPathOfNode);
+    } catch (err) {
+        console.error('ERROR:', err);
+    }
+    const tauriDestFolder = `${__dirname}/../src/node`;
+    await removeDir(tauriDestFolder)
+
+    await copyDir(fullPathOfNode, tauriDestFolder);
+
+}
+
+/**
+ * Copies a directory from the source path to the destination path asynchronously.
+ * @param {string} source - The path of the source directory.
+ * @param {string} destination - The path of the destination directory.
+ * @returns {Promise<void>} - A promise that resolves when the directory is successfully copied, or rejects with an error.
+ * @throws {Error} - If an error occurs during the copying process.
+ */
+async function copyDir(source, destination) {
+    try {
+        console.log(source);
+        console.log(destination);
+        await fsExtra.copy(source, destination);
+        console.log('Successfully copied the folder!');
+    } catch (err) {
+        console.error('An error occurred: ', err);
+    }
+}
+
+/**
+ * Asynchronously removes a specified directory if it exists.
+ *
+ * @async
+ * @function
+ * @param {string} dirPath - The path of the directory to be removed.
+ * @returns {Promise<string>} A promise that resolves to a string indicating whether the directory was not found or was successfully removed.
+ * @throws {Error} If an error occurs during the operation, it will be logged to the console.
+ */
+async function removeDir(dirPath) {
+    try {
+        const exists = await fsExtra.pathExists(dirPath);
+
+        if (!exists) {
+            return 'Directory not found!';
+        }
+
+        await fsExtra.remove(dirPath);
+        return 'Directory removed!';
+    } catch (err) {
+        console.error(err);
+    }
+}
+
+//
+// try {
+//     const version = await fetchLatestNodeVersion();
+//     // const fileName = await downloadNodeBinary(version, 'linux', 'x64');
+//     // await untarFile(fileName, ".");
+//     const fileName = await downloadNodeBinary(version, 'win', 'x64');
+//     unzipFile(fileName, ".");
+//     console.log('Downloaded file:', fileName);
+// } catch (err) {
+//     console.error(err);
+// }
+
+await copyLatestNodeForBuild("linux", "x64");
